Read NK/OT/HT as utf32
Fix wtext color

+++ include/viewer/box_viewer.hpp
@@ -17,9 +17,9 @@ struct vPkm_s
 	bool isInfected : 1;
 	bool isCured : 1;
 	unsigned : 4;
-	u8 NKName[0xD];
-	u8 OTName[0xD];
-	u8 HTName[0xD];
+	uint32_t NKName[0xD];
+	uint32_t OTName[0xD];
+	uint32_t HTName[0xD];
 	const u8* species;
 	const u8* item;
 	const u8* nature;
